pyproject.toml
[project]
name = "adp-sync"
version = "0.1.0"
description = ""
authors = [
  { name = "Charlie Bini", email = "5003326+cbini@users.noreply.github.com" },
]
requires-python = ">=3.7"
license = { text = "GPL-3.0-or-later" }
dependencies = [
  "requests>=2.28.2",
  "google-cloud-storage>=2.7.0",
  "oauthlib>=3.2.2",
  "requests-oauthlib>=1.3.1",
  "pyyaml>=6.0",
]

[project.optional-dependencies]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.pdm.dev-dependencies]
dev = ["pytest"]

[tool.pdm.scripts]
clean-build = { shell = "bash .pdm/scripts/clean-build.sh", help = "remove build artifacts" }
clean-pyc = { shell = "bash .pdm/scripts/clean-pyc.sh", help = "remove Python file artifacts" }
clean-test = { shell = "bash .pdm/scripts/clean-test.sh", help = "remove test and coverage artifacts" }
clean = { shell = "pdm run clean-build && pdm run clean-pyc && pdm run clean-test", help = "remove all build, test, coverage and Python artifacts" }
docs-build = { shell = "mkdocs build", help = "Build the MkDocs documentation" }
docs-serve = { shell = "mkdocs serve", help = "Run the builtin development server" }
docs-deploy = { shell = "mkdocs gh-deploy", help = "Deploy your documentation to GitHub Pages" }
test = { shell = "pdm run pytest", help = "run tests with the default Python" }
extract-workers = { call = "adp_sync.extract:main" }
update-workers = { call = "adp_sync.workers_update:main" }
extract-wfm = { call = "adp_sync.wfm_extract:main" }
post_lock = { shell = "pdm export -f requirements -o requirements.txt" }
